# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareerHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account (seeker or company)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /company/profile:
    get:
      operationId: getCompanyProfile
      tags: [company]
      summary: Get this company's own profile
      responses:
        "200":
          description: Company profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCompanyProfile
      tags: [company]
      summary: Update company profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /company/dashboard:
    get:
      operationId: getCompanyDashboard
      tags: [company]
      summary: Company dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyDashboard"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /company/jobs:
    get:
      operationId: getCompanyJobs
      tags: [company]
      summary: List all jobs posted by this company
      responses:
        "200":
          description: List of company jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobWithApplicantCount"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /company/jobs/{jobId}/applicants:
    get:
      operationId: getJobApplicants
      tags: [company]
      summary: List all applicants for a specific job (company only)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of applicants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationWithSeeker"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /jobs/{jobId}/apply:
    post:
      operationId: applyToJob
      tags: [applications]
      summary: Apply to a job (seeker only)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Already applied or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a seeker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found or closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{applicationId}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status (company only)
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /companies/{companyUserId}/profile:
    get:
      operationId: getPublicCompanyProfile
      tags: [company]
      summary: Get a company's public profile (for job detail views)
      parameters:
        - name: companyUserId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Company profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyProfile"
        "404":
          description: Company not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth schemas
    SignupInput:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [seeker, company]
        full_name:
          type: ["string", "null"]
        headline:
          type: ["string", "null"]
        company_name:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, email, role, created_at]
      properties:
        id:
          type: number
        email:
          type: string
        role:
          type: string
          enum: [seeker, company]
        created_at:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # Seeker schemas
...
    CompanyProfile:
      type: object
      required: [user_id, company_name, industry, company_size, location]
      properties:
        user_id:
          type: number
        company_name:
          type: string
        logo_url:
          type: ["string", "null"]
        industry:
          type: string
        company_size:
          type: string
          enum: ["1-10", "11-50", "51-200", "201-1000", "1000+"]
        website_url:
          type: ["string", "null"]
        about:
          type: ["string", "null"]
        location:
          type: string
    CompanyProfileInput:
      type: object
      properties:
        company_name:
          type: string
        logo_url:
          type: ["string", "null"]
        industry:
          type: string
        company_size:
          type: string
          enum: ["1-10", "11-50", "51-200", "201-1000", "1000+"]
        website_url:
          type: ["string", "null"]
        about:
          type: ["string", "null"]
        location:
          type: string
    CompanyDashboard:
      type: object
      required: [active_postings, total_applicants, awaiting_review]
      properties:
        active_postings:
          type: number
        total_applicants:
          type: number
        awaiting_review:
          type: number

    # Job schemas
...
    JobWithApplicantCount:
      type: object
      required: [id, title, location, work_mode, job_type, experience_level, status, created_at, applicant_count, posted_by]
      properties:
        id:
          type: number
        title:
          type: string
        location:
          type: string
        work_mode:
          type: string
          enum: [On-site, Remote, Hybrid]
        job_type:
          type: string
          enum: [Full-time, Part-time, Contract, Internship]
        experience_level:
          type: string
          enum: ["Entry / Student", Mid, Senior]
        status:
          type: string
          enum: [open, closed]
        created_at:
          type: string
        applicant_count:
          type: number
        posted_by:
          type: number
...
    Application:
      type: object
      required: [id, job_id, applicant_id, status, created_at]
      properties:
        id:
          type: number
        job_id:
          type: number
        applicant_id:
          type: number
        cover_note:
          type: ["string", "null"]
        resume_url:
          type: ["string", "null"]
        status:
          type: string
          enum: [submitted, under_review, shortlisted, rejected, hired]
        created_at:
          type: string
    ApplicationInput:
      type: object
      properties:
        cover_note:
          type: ["string", "null"]
        resume_url:
          type: ["string", "null"]
    ApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [submitted, under_review, shortlisted, rejected, hired]
...
    ApplicationWithSeeker:
      type: object
      required: [id, job_id, applicant_id, status, created_at, seeker_name, seeker_headline]
      properties:
        id:
          type: number
        job_id:
          type: number
        applicant_id:
          type: number
        cover_note:
          type: ["string", "null"]
        resume_url:
          type: ["string", "null"]
        status:
          type: string
          enum: [submitted, under_review, shortlisted, rejected, hired]
        created_at:
          type: string
        seeker_name:
          type: string
        seeker_headline:
          type: ["string", "null"]

    # Stats
...